+++ src/namespace.ts
@@ -11,7 +11,7 @@ export function getNamespace(doc: vscode.TextDocument, position: vscode.Position
   if (outputWindow.isResultsDoc(doc)) {
     const outputWindowNs = outputWindow.getNs();
     utilities.assertIsDefined(outputWindowNs, 'Expected output window to have a namespace!');
-    return outputWindowNs;
+    return [outputWindowNs, `(in-ns '${outputWindowNs})`];
   }
   if (doc && doc.languageId == 'clojure') {
     try {
@@ -20,7 +20,7 @@ export function getNamespace(doc: vscode.TextDocument, position: vscode.Position
         nsUtil.nsFromCursorDoc(
           cursorDoc,
           position ? doc.offsetAt(position) : doc.getText().length
-        ) ?? 'user'
+        ) ?? ['user', "(in-ns 'user)"]
       );
     } catch (e) {
       console.log(
@@ -33,7 +33,7 @@ export function getNamespace(doc: vscode.TextDocument, position: vscode.Position
           if (nsFormArray != undefined && nsFormArray.length > 0) {
             const nsForm = nsFormArray[0].filter((x) => typeof x == 'string');
             if (nsForm != undefined) {
-              return nsForm[1];
+              return [nsForm[1], `(in-ns '${nsForm[1]})`];
             }
           }
         }
@@ -42,7 +42,7 @@ export function getNamespace(doc: vscode.TextDocument, position: vscode.Position
       }
     }
   }
-  return 'user';
+  return ['user', "(in-ns 'user)"];
 }
 
 export async function createNamespaceFromDocumentIfNotExists(doc) {